Bring develop current with release/8.4 via intermediate merge

+++ WordPress/Classes/ViewRelated/Aztec/ViewControllers/AztecPostViewController.swift
@@ -1859,7 +1859,7 @@ extension AztecPostViewController {
 
         let headerOptions = Constants.headers.map { headerType -> OptionsTableViewOption in
             let attributes = [
-                NSFontAttributeName: UIFont.systemFont(ofSize: headerType.fontSize),
+                NSFontAttributeName: UIFont.systemFont(ofSize: CGFloat(headerType.fontSize)),
                 NSForegroundColorAttributeName: WPStyleGuide.darkGrey()
             ]
